Refactor & move CollectionScreen.
